refactor: Change structure of components
